This assignment:
Write a C++ program to allow the user to enter name-score pairs, sort the scores in ascending order and calculates the average score. This program dynamically allocates an array large enough to hold a user-defined number of test scores and also dynamically allocates an array to hold the student names. For each student taking a test, the user types the student’s name followed by the student’s integer test score. Once all the scores are entered, the array holding test scores and the array holding student names should be passed to a function that sorts the scores in ascending order. When the sorted list of scores is displayed, each student’s name should be displayed along with his or her score. Another function should be called that calculates the average score (format the average score with two digits of precision to the right of the decimal point). The output should be aligned.